prevent multiple instances of grit

+++ runscript.sh
@@ -6,5 +6,10 @@ if [ -e ./log/grit.out ]
   rm -f ./log/grit.out
   echo "old grit.old deleted"
 fi
-nohup ./bin/GRIT java -Xdebug -Xnoagent \ -Djava.compiler=NONE \ -Xrunjdwp:transport=dt_socket,server=y,suspend=y,address=6001 &>./log/grit.out &
-echo $! > grit.pid
+if [ -e grit.pid ]
+ then
+  echo "Grit maybe already running, please execute shutdownGrit.sh first."
+ else
+  nohup ./bin/GRIT java -Xdebug -Xnoagent \ -Djava.compiler=NONE \ -Xrunjdwp:transport=dt_socket,server=y,suspend=y,address=6001 &>./log/grit.out &
+  echo $! > grit.pid
+fi
